fix(commonjs): Handle shadowed imports for multiple requires
