+++ packages/commonjs/src/ast-utils.js
@@ -1,10 +1,3 @@
-import { dirname, resolve } from 'path';
-
-import { sync as nodeResolveSync } from 'resolve';
-
-import { DYNAMIC_JSON_PREFIX, DYNAMIC_REGISTER_PREFIX } from './helpers';
-import { normalizePathSlashes } from './utils';
-
 export { default as isReference } from 'is-reference';
 
 const operators = {
@@ -76,34 +69,6 @@ export function isDefineCompiledEsm(node) {
   return false;
 }
 
-export function getAssignedMember(node) {
-  const { left, operator, right } = node.expression;
-  if (operator !== '=' || left.type !== 'MemberExpression') {
-    return null;
-  }
-  let assignedIdentifier;
-  if (left.object.type === 'Identifier') {
-    // exports.foo = ...
-    assignedIdentifier = left.object;
-  } else if (
-    left.object.type === 'MemberExpression' &&
-    left.object.property.type === 'Identifier'
-  ) {
-    // module.exports.foo = ...
-    assignedIdentifier = left.object.property;
-  } else {
-    return null;
-  }
-
-  if (!['module', 'exports'].includes(assignedIdentifier.name)) {
-    return null;
-  }
-
-  const object = left.object ? left.object.name : null;
-  const key = left.property ? left.property.name : null;
-  return { object, key, value: right };
-}
-
 export function getDefinePropertyCallName(node, targetName) {
   const targetNames = targetName.split('.');
 
@@ -140,143 +105,13 @@ export function getDefinePropertyCallName(node, targetName) {
   return { key: key.value, value: valueProperty.value };
 }
 
-export function isRequireStatement(node, scope) {
-  if (!node) return false;
-  if (node.type !== 'CallExpression') return false;
-
-  // Weird case of `require()` or `module.require()` without arguments
-  if (node.arguments.length === 0) return false;
-
-  return isRequireIdentifier(node.callee, scope);
-}
-
-export function isRequireIdentifier(node, scope) {
-  if (!node) return false;
-
-  if (node.type === 'Identifier' && node.name === 'require' /* `require` */) {
-    // `require` is hidden by a variable in local scope
-    if (scope.contains('require')) return false;
-
-    return true;
-  } else if (node.type === 'MemberExpression' /* `[something].[something]` */) {
-    // `module.[something]`
-    if (node.object.type !== 'Identifier' || node.object.name !== 'module') return false;
-
-    // `module` is hidden by a variable in local scope
-    if (scope.contains('module')) return false;
-
-    // `module.require(...)`
-    if (node.property.type !== 'Identifier' || node.property.name !== 'require') return false;
-
-    return true;
-  }
-
-  return false;
-}
-
-export function isStaticRequireStatement(node, scope) {
-  if (!isRequireStatement(node, scope)) return false;
-  return !hasDynamicArguments(node);
-}
-
-function hasDynamicArguments(node) {
-  return (
-    node.arguments.length > 1 ||
-    (node.arguments[0].type !== 'Literal' &&
-      (node.arguments[0].type !== 'TemplateLiteral' || node.arguments[0].expressions.length > 0))
-  );
-}
-
-const reservedMethod = { resolve: true, cache: true, main: true };
-
-export function isNodeRequirePropertyAccess(parent) {
-  return parent && parent.property && reservedMethod[parent.property.name];
-}
-
-export function isIgnoredRequireStatement(requiredNode, ignoreRequire) {
-  return ignoreRequire(requiredNode.arguments[0].value);
-}
-
-function getRequireStringArg(node) {
-  return node.arguments[0].type === 'Literal'
-    ? node.arguments[0].value
-    : node.arguments[0].quasis[0].value.cooked;
-}
-
-function hasDynamicModuleForPath(source, id, dynamicRequireModuleSet) {
-  if (!/^(?:\.{0,2}[/\\]|[A-Za-z]:[/\\])/.test(source)) {
-    try {
-      const resolvedPath = normalizePathSlashes(nodeResolveSync(source, { basedir: dirname(id) }));
-      if (dynamicRequireModuleSet.has(resolvedPath)) {
-        return true;
-      }
-    } catch (ex) {
-      // Probably a node.js internal module
-      return false;
-    }
-
-    return false;
-  }
-
-  for (const attemptExt of ['', '.js', '.json']) {
-    const resolvedPath = normalizePathSlashes(resolve(dirname(id), source + attemptExt));
-    if (dynamicRequireModuleSet.has(resolvedPath)) {
+export function isLocallyShadowed(name, scope) {
+  while (scope.parent) {
+    if (scope.declarations[name]) {
       return true;
     }
+    // eslint-disable-next-line no-param-reassign
+    scope = scope.parent;
   }
-
   return false;
-}
-
-export function shouldUseSimulatedRequire(required, id, dynamicRequireModuleSet) {
-  return (
-    hasDynamicModuleForPath(required.source, id, dynamicRequireModuleSet) &&
-    // We only do `commonjsRequire` for json if it's the `commonjsRegister` call.
-    (required.source.startsWith(DYNAMIC_REGISTER_PREFIX) || !required.source.endsWith('.json'))
-  );
-}
-
-export function getRequireHandlers(id, dynamicRequireModuleSet) {
-  const requiredSources = [];
-  const requiredBySource = Object.create(null);
-  const dynamicRegisterSources = [];
-
-  // TODO Lukas this is assymetric with regard to "usesRequired": Do this manually outside?
-  function getRequired(node, usesRequired) {
-    let sourceId = getRequireStringArg(node);
-    const isDynamicRegister = sourceId.startsWith(DYNAMIC_REGISTER_PREFIX);
-    if (isDynamicRegister) {
-      sourceId = sourceId.substr(DYNAMIC_REGISTER_PREFIX.length);
-    }
-
-    const existing = requiredBySource[sourceId];
-    if (!existing) {
-      const isDynamic = hasDynamicModuleForPath(sourceId, id, dynamicRequireModuleSet);
-      if (isDynamicRegister) {
-        if (sourceId.endsWith('.json')) {
-          sourceId = DYNAMIC_JSON_PREFIX + sourceId;
-        }
-        dynamicRegisterSources.push(sourceId);
-      }
-
-      if (!isDynamic || sourceId.endsWith('.json')) {
-        requiredSources.push(sourceId);
-      }
-
-      requiredBySource[sourceId] = {
-        source: sourceId,
-        name: null,
-        isDynamic,
-        nodesUsingRequired: []
-      };
-    }
-
-    const required = requiredBySource[sourceId];
-    if (usesRequired) {
-      required.nodesUsingRequired.push(node);
-    }
-    return required;
-  }
-
-  return { getRequired, requiredBySource, requiredSources, dynamicRegisterSources };
 }
