+++ packages/commonjs/test/fixtures/function/shadowed-import/_config.js
@@ -1,0 +1,3 @@
+module.exports = {
+  description: 'makes sure imported variables are not shadowed when used in multiple places'
+};
