+++ packages/commonjs/test/fixtures/function/shadowed-import/foo.js
@@ -1,0 +1,1 @@
+module.exports = 'foo';
