+++ packages/commonjs/test/fixtures/function/shadowed-import/main.js
@@ -1,0 +1,12 @@
+const foo = require('./foo');
+
+t.is(foo, 'foo');
+
+{
+  // eslint-disable-next-line no-shadow
+  const foo = 'wrong';
+  // eslint-disable-next-line global-require
+  const bar = require('./foo');
+  t.is(foo, 'wrong');
+  t.is(bar, 'foo');
+}
