+++ packages/commonjs/test/snapshots/function.js.md
@@ -4722,6 +4722,36 @@ Generated by [AVA](https://avajs.dev).
       `,
     }
 
+## shadowed-import
+
+> Snapshot 1
+
+    {
+      'main.js': `'use strict';␊
+      ␊
+      var foo = 'foo';␊
+      ␊
+      const foo$1 = foo;␊
+      ␊
+      t.is(foo$1, 'foo');␊
+      ␊
+      {␊
+        // eslint-disable-next-line no-shadow␊
+        const foo$1 = 'wrong';␊
+        // eslint-disable-next-line global-require␊
+        const bar = foo;␊
+        t.is(foo$1, 'wrong');␊
+        t.is(bar, 'foo');␊
+      }␊
+      ␊
+      var main = {␊
+      ␊
+      };␊
+      ␊
+      module.exports = main;␊
+      `,
+    }
+
 ## shadowing
 
 > Snapshot 1
@@ -5428,6 +5458,33 @@ Generated by [AVA](https://avajs.dev).
       var path__default = /*#__PURE__*/_interopDefaultLegacy(path);␊
       ␊
       t.is(typeof path__default['default'].resolve, 'function');␊
+      ␊
+      var main = {␊
+      ␊
+      };␊
+      ␊
+      module.exports = main;␊
+      `,
+    }
+
+## shadowed-nested-import
+
+> Snapshot 1
+
+    {
+      'main.js': `'use strict';␊
+      ␊
+      var foo = 'foo';␊
+      ␊
+      var bar = 'bar';␊
+      ␊
+      t.is(foo, 'foo');␊
+      ␊
+      {␊
+        // eslint-disable-next-line global-require␊
+        const bar$1 = bar;␊
+        t.is(bar$1, 'bar');␊
+      }␊
       ␊
       var main = {␊
       ␊
